requirements.txt
art==5.8
et-xmlfile==1.1.0
openpyxl==3.0.10
yattag==1.14.0
python-dotenv==0.21.1
pyTelegramBotAPI
requests~=2.28.1